+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- repeatable set conditional logic not evaluating predicate elements correctly
+
 ## [5.1.0] - 2023-11-20
 
 ### Added